Moved Base models in db_utils.py.
Preparing to make everything more modular.
Couldn't put in utils for the moment since utils depends on models and we would get circular imports this way.
After modularizing everything models.py should be gone anyway so we can then merge db_utils.py and utils.py
